+++ blossom/updates/create_leaderboards_table.php
@@ -8,17 +8,26 @@ class CreateLeaderboardsTable extends Migration
 
     public function up()
     {
-        Schema::create('delphinium_blossom_leaderboards', function($table)
+         if ( !Schema::hasTable('delphinium_blossom_leaderboards') )
         {
-            $table->engine = 'InnoDB';
-            $table->increments('id');
-            $table->timestamps();
-        });
+            Schema::create('delphinium_blossom_leaderboards', function($table)
+            {
+                $table->engine = 'InnoDB';
+                $table->increments('id');
+                $table->string('Name');
+                $table->string('Animate');
+                $table->string('Size');
+                $table->timestamps();
+            });
+        }
     }
 
     public function down()
     {
-        Schema::dropIfExists('delphinium_blossom_leaderboards');
+        if ( Schema::hasTable('delphinium_blossom_leaderboards') )
+        {
+            Schema::dropIfExists('delphinium_blossom_leaderboards');
+        }
     }
 
 }
